# server/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Divorce Ledger API
  description: |
    Forensic Financial & Legal Case Management API for divorce proceedings.

    ## Authentication
    Most endpoints require session-based authentication via cookie. Admin endpoints require an `x-admin-secret` header.

    ## Environment Separation
    All data operations are scoped to the user's current environment (LIVE or DEMO).
    DEMO data resets daily at midnight UTC.

    ## Subscription Tiers
    Features are gated based on subscription tier: Free, Individual, Pro, Team, Enterprise.

    ## Rate Limiting
    - Login endpoint: 5 attempts per 15 minutes
    - General API: 100 requests per minute

    ## Version Management
    API versioning follows semantic versioning. Breaking changes increment the major version.
    Current version is included in response headers as `X-API-Version`.
  version: 1.0.0
  contact:
    name: Divorce Ledger Support
    email: support@divorceledger.live
  license:
    name: Proprietary
    url: https://divorceledger.live/license
servers:
  - url: /api
    description: API Base URL (all paths are relative to /api)
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: connect.sid
      description: Session cookie for authenticated users
    adminSecret:
      type: apiKey
      in: header
      name: x-admin-secret
      description: Admin secret for privileged operations
  schemas:
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Human-readable error message
          example: 'Resource not found'
        code:
          type: string
          description: Machine-readable error code
          example: 'NOT_FOUND'
        details:
          type: object
          description: Additional error context
          additionalProperties: true
    ValidationError:
      type: object
      required:
        - error
        - details
      properties:
        error:
          type: string
          example: 'Validation failed'
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: 'email'
              message:
                type: string
                example: 'Invalid email format'
              code:
                type: string
                example: 'invalid_format'
    Pagination:
      type: object
      properties:
        limit:
          type: integer
          minimum: 1
          maximum: 100
          default: 50
          description: Number of items per page
        offset:
          type: integer
          minimum: 0
          default: 0
          description: Number of items to skip
        total:
          type: integer
          description: Total number of items available

    # User schemas
...
    Event:
      type: object
      required:
        - eventId
        - eventType
        - topic
        - payload
        - sequenceNumber
        - createdAt
      properties:
        eventId:
          type: string
        eventType:
          type: string
          description: Event type identifier
        topic:
          type: string
          description: Event topic for routing
        partitionKey:
          type: string
          description: Partition key for ordering
        payload:
          type: object
          additionalProperties: true
        metadata:
          type: object
          additionalProperties: true
        sequenceNumber:
          type: string
        correlationId:
          type: string
        createdAt:
          type: string
          format: date-time
    PublishEventRequest:
      type: object
      required:
        - eventType
        - topic
        - payload
      properties:
        eventType:
          type: string
          minLength: 1
          maxLength: 100
        topic:
          type: string
          minLength: 1
          maxLength: 100
          description: |
            Event topics:
            - user.tier_changed
            - billing.payment
            - violation.created
            - evidence.uploaded
            - ai.classification
        payload:
          type: object
          additionalProperties: true
        partitionKey:
          type: string
          maxLength: 100
        correlationId:
          type: string
          maxLength: 100
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: 'Authentication required'
            code: 'UNAUTHORIZED'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: 'Resource not found'
            code: 'NOT_FOUND'
    ValidationFailed:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
...
paths:
...
  /events/status:
    get:
      tags: [Events]
      summary: Get event streaming status
      description: Returns processor status, topic stats, and dead letter queue count
      security:
        - adminSecret: []
      responses:
        '200':
          description: Event streaming status
          content:
            application/json:
              schema:
                type: object
                properties:
                  processors:
                    type: object
                  topics:
                    type: array
                    items:
                      type: object
                  outbox:
                    type: object
                  deadLetterQueue:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
  /events/publish:
    post:
      tags: [Events]
      summary: Publish an event
      description: Publishes an event to the event streaming system
      security:
        - adminSecret: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PublishEventRequest'
      responses:
        '200':
          description: Event published
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  event:
                    $ref: '#/components/schemas/Event'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationFailed'
...
  /events:
    get:
      tags: [Events]
      summary: List events
      description: Returns paginated list of events with optional topic filter
      security:
        - adminSecret: []
      parameters:
        - name: topic
          in: query
          schema:
            type: string
          description: Filter by topic
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of events
          content:
            application/json:
              schema:
                type: object
                properties:
                  events:
                    type: array
                    items:
                      $ref: '#/components/schemas/Event'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
  /events/topics:
    get:
      tags: [Events]
      summary: List event topics
      description: Returns all topics with event counts
      security:
        - adminSecret: []
      responses:
        '200':
          description: Topic statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  topics:
                    type: array
                    items:
                      type: object
                      properties:
                        topic:
                          type: string
                        totalEvents:
                          type: integer
                        partitions:
                          type: integer
                        firstEvent:
                          type: string
                          format: date-time
                        lastEvent:
                          type: string
                          format: date-time
  /events/consumers:
    get:
      tags: [Events]
      summary: List consumer groups
      description: Returns all consumer groups with offset lag
      security:
        - adminSecret: []
      responses:
        '200':
          description: Consumer groups
          content:
            application/json:
              schema:
                type: object
                properties:
                  consumerGroups:
                    type: array
                    items:
                      type: object
                      properties:
                        groupId:
                          type: string
                        offsets:
                          type: array
                          items:
                            type: object
                            properties:
                              topic:
                                type: string
                              lastOffset:
                                type: integer
                              lag:
                                type: integer
  /events/dlq:
    get:
      tags: [Events]
      summary: List dead letter queue
      description: Returns failed events in the dead letter queue
      security:
        - adminSecret: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Dead letter queue entries
          content:
            application/json:
              schema:
                type: object
                properties:
                  deadLetterQueue:
                    type: array
                    items:
                      type: object
                      properties:
                        dlqId:
                          type: string
                        eventType:
                          type: string
                        topic:
                          type: string
                        payload:
                          type: object
                        error:
                          type: string
                        retryCount:
                          type: integer
                        createdAt:
                          type: string
                          format: date-time
  /events/dlq/{id}/retry:
    post:
      tags: [Events]
      summary: Retry dead letter event
      description: Republishes a failed event from the dead letter queue
      security:
        - adminSecret: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Event retried
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  dlqId:
                    type: string
        '404':
          $ref: '#/components/responses/NotFound'
  /events/processors/start:
    post:
      tags: [Events]
      summary: Start event processors
      description: Starts analytics and audit event processors
      security:
        - adminSecret: []
      responses:
        '200':
          description: Processors started
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  status:
                    type: object
  /events/processors/stop:
    post:
      tags: [Events]
      summary: Stop event processors
      description: Stops all running event processors
      security:
        - adminSecret: []
      responses:
        '200':
          description: Processors stopped
  /events/consumers/{groupId}/reset:
    post:
      tags: [Events]
      summary: Reset consumer offset
      description: Resets consumer group offset for a topic
      security:
        - adminSecret: []
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - topic
              properties:
                topic:
                  type: string
                toEventId:
                  type: integer
                  default: 0
      responses:
        '200':
          description: Offset reset
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  consumerGroup:
                    type: string
                  topic:
                    type: string
                  newOffset:
                    type: integer

  # ETL Pipeline Endpoints
...